tests/conftest.py
import pytest
from dotenv import load_dotenv


@pytest.fixture(scope="session", autouse=True)
def load_env():
    """Load environment variables from .env file before tests are run."""
    load_dotenv()
